move method to risk model
