+++ AngularApp/projects/app-service-diagnostics/src/app/home/components/risk-tile/risk-tile.component.html
@@ -10,7 +10,7 @@
   </div>
 
   <div class="risk-section">
-    <a>{{link}}</a>
+    <a>{{linkText}}</a>
   </div>
 </div>
